chore : extension 수정
